backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2
django-cors-headers
django-colorfield